+++ release-notes.md
@@ -1,3 +1,7 @@
+`0.2.3` - ???
+* Fix for non-working adversary status to apply
+* NEW FEATURE: Add ability to show minion group (max) size
+
 `0.2.2` - 2021-11-05
 * Fix for relative URLs (thanks `anthonyscorrea`!) ([#68](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/68))
 * Fix for actor renaming resetting on reload (thanks `@Bolts#9418`!)